Move BebMock into testing package
